@@ -1,12 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.3'
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 23
-        targetSdkVersion 25
+        minSdkVersion 19
+        targetSdkVersion 26
         versionCode 35
         versionName "1.2.1"
     }
@@ -33,22 +32,21 @@ android {
 
 repositories {
     mavenCentral()
-    maven { url "https://jitpack.io" }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:recyclerview-v7:25.3.1'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile 'com.android.support:support-v4:25.3.1'
-    compile 'commons-io:commons-io:2.4'
-    compile 'com.github.navasmdc:MaterialDesign:1.5@aar'
-    compile('com.github.afollestad.material-dialogs:core:0.8.5.3@aar') { transitive = true; }
-    compile('com.github.afollestad.material-dialogs:commons:0.8.5.3@aar') { transitive = true; }
-    compile('com.mikepenz:materialdrawer:4.6.1@aar') { transitive = true; }
-    compile 'com.mikepenz:google-material-typeface:1.2.0.1@aar'
-    compile 'com.jaredrummler:colorpicker:1.0.0'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.google.code.gson:gson:2.8.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'commons-io:commons-io:2.4'
+    implementation 'com.github.navasmdc:MaterialDesignLibrary:master-SNAPSHOT'
+    implementation('com.github.afollestad.material-dialogs:core:0.9.6.0@aar') { transitive = true; }
+    implementation('com.github.afollestad.material-dialogs:commons:0.9.6.0@aar') { transitive = true; }
+    implementation('com.mikepenz:materialdrawer:4.6.1@aar') { transitive = true; }
+    implementation 'com.mikepenz:google-material-typeface:1.2.0.1@aar'
+    implementation 'com.jaredrummler:colorpicker:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'com.google.code.gson:gson:2.8.5'
 }
