@@ -5,8 +5,8 @@ import android.app.ActivityOptions;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.drawable.BitmapDrawable;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -21,8 +21,6 @@ import com.dkanada.openapk.models.AppItem;
 import com.dkanada.openapk.utils.Actions;
 import com.dkanada.openapk.activities.AppActivity;
 import com.dkanada.openapk.R;
-import com.dkanada.openapk.utils.FileOperations;
-import com.dkanada.openapk.utils.OtherUtils;
 
 import java.util.ArrayList;
 import java.util.List;
