@@ -1,7 +1,7 @@
 package com.dkanada.openapk.views;
 
 import android.content.Context;
-import android.support.constraint.ConstraintLayout;
+import androidx.constraintlayout.widget.ConstraintLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.LinearLayout;
